Relocate and simplify tests
