+++ tests/test_gp.py
@@ -6,7 +6,7 @@ import pandas as pd
 import pytest
 import torch
 from botorch.fit import fit_gpytorch_mll
-from botorch.models import MultiTaskGP, SingleTaskGP
+from botorch.models import MultiTaskGP, SingleTaskGP, SingleTaskMultiFidelityGP
 from botorch.models.transforms import Normalize, Standardize
 from gpytorch.kernels import MaternKernel as GPyTorchMaternKernel
 from gpytorch.kernels import RBFKernel as GPyTorchRBFKernel
@@ -20,11 +20,18 @@ from pandas.testing import assert_frame_equal
 from pytest import param
 
 from baybe import active_settings
-from baybe.exceptions import ModelNotTrainedError
+from baybe.exceptions import IncompatibleSurrogateError, ModelNotTrainedError
 from baybe.kernels.basic import MaternKernel, RBFKernel
 from baybe.kernels.composite import ScaleKernel
 from baybe.parameters.categorical import TaskParameter
-from baybe.parameters.numerical import NumericalContinuousParameter
+from baybe.parameters.fidelity import (
+    CategoricalFidelityParameter,
+    NumericalDiscreteFidelityParameter,
+)
+from baybe.parameters.numerical import (
+    NumericalContinuousParameter,
+    NumericalDiscreteParameter,
+)
 from baybe.searchspace.core import SearchSpace
 from baybe.surrogates.gaussian_process.components.fit_criterion import FitCriterion
 from baybe.surrogates.gaussian_process.components.generic import PlainGPComponentFactory
@@ -53,6 +60,23 @@ measurements_mt = create_fake_input(
 )
 baybe_kernel = ScaleKernel(MaternKernel() + RBFKernel())
 gpytorch_kernel = GPyTorchScaleKernel(GPyTorchMaternKernel() + GPyTorchRBFKernel())
+
+_num_fid_param = NumericalDiscreteFidelityParameter(
+    "fidelity", values=[0.5, 1.0], costs=[1.0, 10.0]
+)
+_cat_fid_param = CategoricalFidelityParameter(
+    "fidelity", values=["lo", "hi"], costs=[1.0, 10.0], zeta=[0.5, 0.0]
+)
+_design_param = NumericalDiscreteParameter("x", values=[1.0, 2.0, 3.0])
+
+searchspace_num_fid = SearchSpace.from_product([_design_param, _num_fid_param])
+searchspace_cat_fid = SearchSpace.from_product([_design_param, _cat_fid_param])
+measurements_num_fid = create_fake_input(
+    searchspace_num_fid.parameters, objective.targets, n_rows=3
+)
+measurements_cat_fid = create_fake_input(
+    searchspace_cat_fid.parameters, objective.targets, n_rows=3
+)
 
 
 def _dummy_mean_factory(*args, **kwargs) -> GPyTorchMean:
@@ -262,3 +286,90 @@ def test_to_botorch_before_fit():
     """Attempting to access an untrained surrogate raises an error."""
     with pytest.raises(ModelNotTrainedError):
         GaussianProcessSurrogate().to_botorch()
+
+
+@pytest.mark.parametrize(
+    "parameters",
+    [
+        param([TaskParameter("task", values=["a", "b"])], id="task_only"),
+        param([_cat_fid_param], id="categorical_fidelity_only"),
+        param([_num_fid_param], id="numerical_fidelity_only"),
+    ],
+)
+def test_surrogate_rejects_index_only_searchspace(parameters):
+    """GP surrogates raise for search spaces without regular model inputs."""
+    searchspace = SearchSpace.from_product(parameters)
+    measurements = create_fake_input(
+        searchspace.parameters, objective.targets, n_rows=3
+    )
+
+    with pytest.raises(IncompatibleSurrogateError, match="at least one regular"):
+        GaussianProcessSurrogate().fit(searchspace, objective, measurements)
+
+
+def test_gp_rejects_custom_components_numerical_fidelity():
+    """Custom components are rejected for numerical multi-fidelity search spaces."""
+    surrogate = GaussianProcessSurrogate(likelihood_or_factory=GaussianLikelihood())
+    with pytest.raises(IncompatibleSurrogateError, match="custom components"):
+        surrogate.fit(searchspace_num_fid, objective, measurements_num_fid)
+
+
+@pytest.mark.parametrize(
+    ("searchspace", "measurements", "expected_model"),
+    [
+        param(
+            searchspace_cat_fid, measurements_cat_fid, SingleTaskGP, id="categorical"
+        ),
+        param(
+            searchspace_num_fid,
+            measurements_num_fid,
+            SingleTaskMultiFidelityGP,
+            id="numerical",
+        ),
+    ],
+)
+def test_standard_gp_fit_fidelity(searchspace, measurements, expected_model):
+    """GaussianProcessSurrogate fits a fidelity space with the expected model."""
+    surrogate = GaussianProcessSurrogate()
+    surrogate.fit(searchspace, objective, measurements)
+    # Exact type check: SingleTaskMultiFidelityGP subclasses SingleTaskGP, so the
+    # categorical case must not accidentally match the multi-fidelity model.
+    assert type(surrogate.to_botorch()) is expected_model
+    stats = surrogate.posterior_stats(measurements)
+    assert set(stats.columns) == {"t_mean", "t_std"}
+    assert len(stats) == len(measurements)
+
+
+@pytest.mark.parametrize(
+    "preset",
+    [
+        param(
+            preset,
+            marks=pytest.mark.skipif(
+                preset is GaussianProcessPreset.BOTORCH and sys.version_info < (3, 11),
+                reason="BoTorch >=0.18.0 requires Python >=3.11.",
+            ),
+        )
+        for preset in GaussianProcessPreset
+    ],
+    ids=lambda preset: preset.value,
+)
+def test_gp_presets_fit_categorical_fidelity(preset):
+    """All GP presets can be fitted on a categorical fidelity space."""
+    surrogate = GaussianProcessSurrogate.from_preset(preset)
+    surrogate.fit(searchspace_cat_fid, objective, measurements_cat_fid)
+
+
+@pytest.mark.parametrize(
+    "preset", list(GaussianProcessPreset), ids=lambda preset: preset.value
+)
+def test_gp_presets_reject_numerical_fidelity(preset):
+    """No GP preset can be fitted on a numerical fidelity space.
+
+    Numerical multi-fidelity spaces are delegated to BoTorch's
+    ``SingleTaskMultiFidelityGP``, which builds its own components. Since every
+    preset supplies explicit component factories, fitting must be rejected.
+    """
+    surrogate = GaussianProcessSurrogate.from_preset(preset)
+    with pytest.raises(IncompatibleSurrogateError, match="custom components"):
+        surrogate.fit(searchspace_num_fid, objective, measurements_num_fid)
